Separate the composer clipboard wiring test from its behavioral coverage
The structural-test budget charges every case in a test file to the
structural ceiling as soon as that file reads source anywhere. One
source-reading wiring assertion at the end of markdownClipboard.test.js
therefore charged all of its behavioral module tests as implementation-text
debt, so the two new undo-stack tests pushed the count past the ceiling.

Move the wiring assertions into their own file and relax the new one to
match the call rather than its exact argument names. Coverage is unchanged
and the measured structural debt for this area drops from 9 cases to 1.
